feat(generate): 实现代码生成核心逻辑
- 新增 GenCodeCoreBo 类作为代码生成核心业务对象
- 在 GenCodeCoreRequest 中添加 packageName 字段
- 实现代码生成核心服务逻辑，包括：
  - 获取表信息和列信息
  - 解析模板列表
  - 生成 Velocity 上下文
  - 使用 Velocity 模板引擎生成代码
- 新增 GenCodeHelper 工具类，用于辅助代码生成
- 优化 SysMenuEntity 类，移除不必要的 IdType 注解

+++ xht-modules/xht-module-generate/src/main/java/com/xht/generate/domain/entity/GenLogEntity.java
@@ -47,7 +47,7 @@ public class GenLogEntity extends BasicEntity implements Serializable {
      * 生成文件数量
      */
     @TableField(value = "file_count")
-    private Long fileCount;
+    private int fileCount;
 
     /**
      * 生成的表ID（逗号分隔）
